+++ greedy/src/greedy.h
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2024 Yuki Tsuboi
  * 
- * File: greedy.h
+ * File: greedy.cpp
  */
 
 #ifndef GREEDY_SRC_GREEDY_H_
@@ -13,13 +13,14 @@
 #include <algorithm>
 #include <limits>
 #include <sstream>
+#include <cstdint>  // For int64_t
 
-constexpr int MAX_COIN_QUANTITY = 1000000000;
-constexpr int MAX_AMOUNT = 1000000000;
+constexpr int64_t MAX_COIN_QUANTITY = 1000000000;
+constexpr int64_t MAX_AMOUNT = 1000000000;
 
 struct Coin {
     int value;
-    long long quantity;
+    int64_t quantity;  // Changed from long to int64_t
 };
 
 void InputValidation(const std::vector<Coin>& coins, const int& a);
